refactor: [#35] :recycle: Refactor code after rebasing
