+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         <activity android:name=".color.ColorsActivity" />
         <activity android:name=".result.ResultActivity" />
         <activity android:name=".question.QuestionActivity" />
-        <activity android:name=".OptionActivity"></activity>
+        <activity android:name=".option.OptionActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
