+++ app/src/main/java/fr/esgi/rpa/cgg/MainActivity.kt
@@ -7,6 +7,7 @@ import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import fr.esgi.rpa.cgg.color.ColorsActivity
+import fr.esgi.rpa.cgg.option.OptionActivity
 import fr.esgi.rpa.cgg.question.QuestionActivity
 import kotlinx.android.synthetic.main.activity_main.*
 
@@ -15,27 +16,24 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         this.setClickListeners()
-        setOptionButtonClickListener()
-        val opt = initOptionSharedPreferences()
-        // TODO: remove logs
-        opt.getString(OptionActivity.DIFFICULTY_KEY, null)?.let { Log.d("toto", it) }
+
+        val preferences = initOptionSharedPreferences()
+        val difficultyOption = preferences.getString(OptionActivity.DIFFICULTY_KEY, null)
+        Log.v("MainActivity", "DIFFICULTY_KEY = $difficultyOption")
     }
 
+    private fun initOptionSharedPreferences(): SharedPreferences =
+        getSharedPreferences(getString(R.string.optionPreference), Context.MODE_PRIVATE)
+
     private fun setClickListeners() {
-        setPlayButtonClickListener()
-        setColorsButtonClickListener()
+        this.setPlayButtonClickListener()
+        this.setColorsButtonClickListener()
+        this.setOptionButtonClickListener()
     }
 
     private fun setColorsButtonClickListener() {
         colors_button?.setOnClickListener {
             val intent = Intent(this, ColorsActivity::class.java)
-            startActivity(intent)
-        }
-    }
-
-    private fun setPlayButtonClickListener() {
-        play_button?.setOnClickListener {
-            val intent = Intent(this, QuestionActivity::class.java)
             startActivity(intent)
         }
     }
@@ -47,10 +45,10 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun initOptionSharedPreferences(): SharedPreferences {
-        return this.getSharedPreferences(
-            this.getString(R.string.optionPreference),
-            Context.MODE_PRIVATE
-        )
+    private fun setPlayButtonClickListener() {
+        play_button?.setOnClickListener {
+            val intent = Intent(this, QuestionActivity::class.java)
+            startActivity(intent)
+        }
     }
 }
