+++ app/src/main/java/fr/esgi/rpa/cgg/option/OptionActivity.kt
@@ -1,17 +1,15 @@
-package fr.esgi.rpa.cgg
+package fr.esgi.rpa.cgg.option
 
 import android.content.Context
 import android.content.Intent
-import android.content.SharedPreferences
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.widget.Button
+import fr.esgi.rpa.cgg.MainActivity
+import fr.esgi.rpa.cgg.R
 import kotlinx.android.synthetic.main.activity_option.*
 
-
 class OptionActivity : AppCompatActivity() {
-
     private var difficulty: String? = ""
 
     companion object {
@@ -23,24 +21,48 @@ class OptionActivity : AppCompatActivity() {
         setContentView(R.layout.activity_option)
         this.setActualDifficulty()
         this.setClickListeners()
-
     }
 
     private fun setActualDifficulty() {
-        val optionSharedPreferences = this?.getSharedPreferences(this.getString(R.string.optionPreference), Context.MODE_PRIVATE)
+        val optionSharedPreferences =
+            getSharedPreferences(getString(R.string.optionPreference), Context.MODE_PRIVATE)
         difficulty = optionSharedPreferences.getString(DIFFICULTY_KEY, "EASY")
-        when(difficulty) {
+        when (difficulty) {
             "EASY" -> this.setButtonToFocusedState(easy_button, R.drawable.rounded_corners_green)
-            "MEDIUM" -> this.setButtonToFocusedState(medium_button, R.drawable.rounded_corners_yellow)
+            "MEDIUM" -> this.setButtonToFocusedState(
+                medium_button,
+                R.drawable.rounded_corners_yellow
+            )
             "HARD" -> this.setButtonToFocusedState(hard_button, R.drawable.rounded_corners_red)
         }
+    }
+
+    private fun setButtonToDefaultState(button: Button?) {
+        button?.isFocusable = false
+        button?.setBackgroundResource(R.drawable.rounded_corners_primary)
+    }
+
+    private fun setButtonToFocusedState(button: Button?, imageId: Int) {
+        button?.isFocusable = true
+        button?.setBackgroundResource(imageId)
     }
 
     private fun setClickListeners() {
         this.setConfirmButtonClickListener()
         this.setEasyButtonClickListener()
-        this.setMediumButtonClickListener()
         this.setHardButtonClickListener()
+        this.setMediumButtonClickListener()
+    }
+
+    private fun setConfirmButtonClickListener(): Unit? = confirm_button?.setOnClickListener {
+        val optionSharedPreferences =
+            getSharedPreferences(getString(R.string.optionPreference), Context.MODE_PRIVATE)
+        with(optionSharedPreferences.edit()) {
+            putString(DIFFICULTY_KEY, difficulty)
+            apply()
+        }
+        val intent = Intent(this, MainActivity::class.java)
+        this.navigateUpTo(intent)
     }
 
     private fun setEasyButtonClickListener(): Unit? = easy_button?.setOnClickListener {
@@ -50,13 +72,6 @@ class OptionActivity : AppCompatActivity() {
         difficulty = "EASY"
     }
 
-    private fun setMediumButtonClickListener(): Unit? = medium_button?.setOnClickListener {
-        this.setButtonToFocusedState(medium_button, R.drawable.rounded_corners_yellow)
-        this.setButtonToDefaultState(easy_button)
-        this.setButtonToDefaultState(hard_button)
-        difficulty = "MEDIUM"
-    }
-
     private fun setHardButtonClickListener(): Unit? = hard_button?.setOnClickListener {
         this.setButtonToFocusedState(hard_button, R.drawable.rounded_corners_red)
         this.setButtonToDefaultState(easy_button)
@@ -64,24 +79,11 @@ class OptionActivity : AppCompatActivity() {
         difficulty = "HARD"
     }
 
-    private fun setButtonToFocusedState(button: Button, imageId: Int) {
-        button?.isFocusable = true
-        button?.setBackgroundResource(imageId)
-    }
-
-    private fun setButtonToDefaultState(button: Button) {
-        button?.isFocusable = false
-        button?.setBackgroundResource(R.drawable.rounded_corners_primary)
-    }
-
-    private fun setConfirmButtonClickListener(): Unit? = confirm_button?.setOnClickListener {
-        val optionSharedPreferences = this?.getSharedPreferences(this.getString(R.string.optionPreference), Context.MODE_PRIVATE)
-        with(optionSharedPreferences.edit()) {
-            putString(DIFFICULTY_KEY, difficulty)
-            apply()
-        }
-        val intent = Intent(this, MainActivity::class.java)
-        this.navigateUpTo(intent)
+    private fun setMediumButtonClickListener(): Unit? = medium_button?.setOnClickListener {
+        this.setButtonToFocusedState(medium_button, R.drawable.rounded_corners_yellow)
+        this.setButtonToDefaultState(easy_button)
+        this.setButtonToDefaultState(hard_button)
+        difficulty = "MEDIUM"
     }
 
 }
